Fixed command errors and added mask system

+++ server/main.lua
@@ -2,7 +2,6 @@ local players = {}
 -- { identifier }
 
 RegisterServerEvent('map-alias:fetchAlias')
-RegisterServerEvent('map-alias:setServerIdentifier')
 RegisterServerEvent('map-alias:getPlayers')
 
 AddEventHandler('map-alias:fetchAlias', function()
@@ -12,21 +11,15 @@ end)
 function fetchAlias(source)
     local identifier = GetPlayerIdentifier(source)
     local _source = source
+    
     MySQL.query('SELECT * FROM `alias` WHERE identifier = ?', { identifier }, function(result)
         TriggerClientEvent('map-alias:setAlias', _source, result)
-
         for k,_ in pairs(GetPlayers()) do
-            players[k] = {}
+            players[k] = {
+                identifier = GetPlayerIdentifier(k, 0)
+            }
         end
-    
-        getIdentifiers()
     end)
-end
-
-function getIdentifiers()
-    for k,v in pairs(players) do
-        players[k].identifier = GetPlayerIdentifier(k)
-    end
 end
 
 AddEventHandler('map-alias:getPlayers', function()
@@ -35,10 +28,9 @@ end)
 
 RegisterCommand('alias', function(source, args)
     local _source = source
-    if _source > 0 then
-        local localPlayer = GetPlayerIdentifier(_source)
-        local target = GetPlayerIdentifier(args[1])
-        if args[2] == nil then return end
+    if _source > 0 and args[1] and args[2] then
+        local localPlayer = GetPlayerIdentifier(_source, 0)
+        local target = GetPlayerIdentifier(args[1], 0)
 
         if target ~= nil then
             local res = MySQL.query.await('SELECT * FROM `alias` WHERE `target` = ?', { target })
@@ -59,8 +51,8 @@ end)
 
 RegisterCommand('removealias', function(source, args)
     local _source = source
-    if _source > 0 then
-        local target = GetPlayerIdentifier(args[1])
+    if _source > 0 and args[1] then
+        local target = GetPlayerIdentifier(args[1], 0)
         if target ~= nil then
             MySQL.update('DELETE FROM `alias` WHERE `target` = ?', { target }, function()
                 notify(_source, Config.Locale["aliasRemoved"])
